requirements.txt
asgiref==3.3.4
boto3==1.17.72
botocore==1.20.72
dj-database-url==0.5.0
Django==3.2.2
django-heroku==0.3.1
django-storages==1.11.1
django-summernote==0.8.11.6
django-taggit==1.4.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.2.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
